refactor: domain에서 spring 의존성 제거
