+++ src/main/java/racingcar/domain/Cars.java
@@ -4,7 +4,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
-import racingcar.NumberGenerator;
 
 public class Cars {
