+++ src/main/java/racingcar/domain/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface NumberGenerator {
     int generate();
