+++ src/main/java/racingcar/domain/RandomNumberGenerator.java
@@ -1,8 +1,5 @@
-package racingcar;
+package racingcar.domain;
 
-import org.springframework.stereotype.Component;
-
-@Component
 public class RandomNumberGenerator implements NumberGenerator {
     public static final int BOUND = 10;
