+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -10,7 +10,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.NumberGenerator;
 
 class RacingGameTest {
     private final AlwaysMoveNumberGenerator alwaysMoveNumberGenerator = new AlwaysMoveNumberGenerator();
